# documentation/openapi.json
{"openapi":"3.0.2","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/status":{"get":{"tags":["Status"],"summary":"A health check for the API.","operationId":"health_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StatusServer"}}}}}}},"/register":{"post":{"tags":["Authoriation"],"summary":"To register a customer","description":"Only registers a customer (not a merchant).\n\nEncodes password in the SHA256 encoding standard.\n\nReturns the JWT token (calculated with the HS256 algorithm) with expiry of 1 week.","operationId":"register_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterData"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthReponse"}}}},"401":{"description":"A user with the given email is already registered"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth":{"post":{"tags":["Authoriation"],"summary":"To login a user","description":"Valid for either merchants or customers.\nCompares the password based on the SHA256 algorithm.\nReturns the JWT token (calculated with the HS256 algorithm) with expiry of 1 week.","operationId":"login_auth_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginData"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthReponse"}}}},"401":{"description":"The given email does not exist or the password doesn't match"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/me/det":{"get":{"tags":["User Details"],"summary":"To get the user details","operationId":"get_user_me_det_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDet"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/me/hasOngoing":{"get":{"tags":["User Details"],"summary":"To check if the user has an ongoing order","description":"Checks the database if the user has at least 1 order where is_complete=0","operationId":"ongoing_me_hasOngoing_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDet"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/getStores":{"get":{"tags":["Store Views"],"summary":"Gets all the stores in the database","description":"Configures the start and end of each store to convert from UNIX time to user-friendly times.","operationId":"stores_getStores_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Stores Getstores Get","type":"array","items":{"$ref":"#/components/schemas/Store"}}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/getStoreItems":{"get":{"tags":["Store Views"],"summary":"Gets all the stores items according to category and pagination variables","description":"Combines the store items with the user's cart items to get the quantity of each item.","operationId":"storeItems_getStoreItems_get","parameters":[{"required":true,"schema":{"title":"The id of the store","type":"integer"},"name":"store_id","in":"query"},{"required":true,"schema":{"title":"The category of items to get","type":"string"},"name":"category","in":"query"},{"required":true,"schema":{"title":"The items to skip over (for pagination).","type":"integer"},"name":"skip","in":"query"},{"required":false,"schema":{"title":"The number of results per page (for pagination)","type":"integer","default":8},"name":"resPerPage","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemsResp"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/getStoreCategories":{"get":{"tags":["Store Views"],"summary":"A list of the store's categories","operationId":"storeCategories_getStoreCategories_get","parameters":[{"required":true,"schema":{"title":"The id of the store","type":"integer"},"name":"store_id","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Storecategories Getstorecategories Get","type":"array","items":{"type":"string"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/getRecommendedItems":{"get":{"tags":["Recommendations"],"summary":"Recitems","operationId":"recItems_getRecommendedItems_get","parameters":[{"required":true,"schema":{"title":"The id of the store","type":"integer"},"name":"store_id","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/getRecommendedStores":{"get":{"tags":["Recommendations"],"summary":"Recitems","operationId":"recItems_getRecommendedStores_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/storeAvailability":{"get":{"tags":["Availability"],"summary":"Availability","operationId":"availability_storeAvailability_get","parameters":[{"required":true,"schema":{"title":"Store Id","type":"integer"},"name":"store_id","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/modifyCart":{"put":{"tags":["Cart"],"summary":"Modifycart","operationId":"modifyCart_modifyCart_put","parameters":[{"required":true,"schema":{"title":"Store Id","type":"integer"},"name":"store_id","in":"query"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_modifyCart_modifyCart_put"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/myLatestCart":{"get":{"tags":["Cart"],"summary":"Usercart","operationId":"userCart_myLatestCart_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/submitOrder":{"post":{"tags":["Order"],"summary":"Submitorder","operationId":"submitOrder_submitOrder_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_submitOrder_submitOrder_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/myLatestOrder":{"get":{"tags":["Order"],"summary":"Getorder","operationId":"getOrder_myLatestOrder_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/orderParking":{"post":{"tags":["Order"],"summary":"Orderparking","operationId":"orderParking_orderParking_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_orderParking_orderParking_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/orders":{"get":{"tags":["Order"],"summary":"Get Orders","operationId":"get_orders_orders_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/markComplete":{"put":{"tags":["Order"],"summary":"Markcomplete","operationId":"markComplete_markComplete_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_markComplete_markComplete_put"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/storeSettings":{"get":{"summary":"Getstoresettings","operationId":"getStoreSettings_storeSettings_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/storeFrequency":{"post":{"tags":["Availability"],"summary":"Storefreq","operationId":"storeFreq_storeFrequency_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_storeFreq_storeFrequency_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/storeCapacity":{"post":{"tags":["Availability"],"summary":"Storecapacity","operationId":"storeCapacity_storeCapacity_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_storeCapacity_storeCapacity_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}}},"components":{"schemas":{"AuthReponse":{"title":"AuthReponse","required":["access_token","role"],"type":"object","properties":{"access_token":{"title":"The JWT access token upon user registration/login","type":"string"},"role":{"title":"The role of the user. Either CUSTOMER or MERCHANT. Used for page redirection on UI.","type":"string"}},"example":{"access_token":"[JWT ACCESS TOKEN HERE]","role":"CUSTOMER"}},"Body_markComplete_markComplete_put":{"title":"Body_markComplete_markComplete_put","required":["order_id"],"type":"object","properties":{"order_id":{"title":"Order Id","type":"integer"}}},"Body_modifyCart_modifyCart_put":{"title":"Body_modifyCart_modifyCart_put","required":["itemId","quant"],"type":"object","properties":{"itemId":{"title":"Itemid","type":"integer"},"quant":{"title":"Quant","type":"integer"}}},"Body_orderParking_orderParking_post":{"title":"Body_orderParking_orderParking_post","required":["parkingNum"],"type":"object","properties":{"parkingNum":{"title":"Parkingnum","type":"integer"}}},"Body_storeCapacity_storeCapacity_post":{"title":"Body_storeCapacity_storeCapacity_post","required":["slotCapacity"],"type":"object","properties":{"slotCapacity":{"title":"Slotcapacity","type":"integer"}}},"Body_storeFreq_storeFrequency_post":{"title":"Body_storeFreq_storeFrequency_post","required":["slotFrequency"],"type":"object","properties":{"slotFrequency":{"title":"Slotfrequency","type":"integer"}}},"Body_submitOrder_submitOrder_post":{"title":"Body_submitOrder_submitOrder_post","required":["time"],"type":"object","properties":{"time":{"title":"Time","type":"string"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"ItemsResp":{"title":"ItemsResp","required":["pageCount","items"],"type":"object","properties":{"pageCount":{"title":"The number of pages for this category","type":"integer"},"items":{"title":"Items","type":"array","items":{"$ref":"#/components/schemas/StoreItem"}}}},"LoginData":{"title":"LoginData","required":["email","password"],"type":"object","properties":{"email":{"title":"The user's email","type":"string"},"password":{"title":"The user's password","type":"string"}},"example":{"email":"example@gmail.com","password":"test"}},"RegisterData":{"title":"RegisterData","required":["email","password","first_name","last_name","phone_number"],"type":"object","properties":{"email":{"title":"The user's email","type":"string"},"password":{"title":"The user's password","type":"string"},"first_name":{"title":"The user's first name","type":"string"},"last_name":{"title":"The user's last name","type":"string"},"phone_number":{"title":"The user's phone number","type":"string"}},"example":{"email":"example@gmail.com","password":"test","first_name":"Joe","last_name":"test","phone_number":"8326910295"}},"StatusServer":{"title":"StatusServer","type":"object","properties":{"status":{"title":"If the server is online or not","type":"string"}},"example":{"status":"online"}},"Store":{"title":"Store","required":["id","name","lat","lng","desc","image","start","end"],"type":"object","properties":{"id":{"title":"The ID of the store","type":"integer"},"name":{"title":"The name of the store","type":"string"},"lat":{"title":"The latitude (coordinates) of the store","type":"number"},"lng":{"title":"The longitude of the store","type":"number"},"desc":{"title":"The description of the store","type":"string"},"image":{"title":"The relative link to the logo of the store","type":"string"},"start":{"title":"The opening hour of the store. Format: HH:MM AM/PM","type":"string"},"end":{"title":"The closing hour of the store. Format: HH:MM AM/PM","type":"string"}}},"StoreItem":{"title":"StoreItem","required":["id","name","price","image","description","quantity"],"type":"object","properties":{"id":{"title":"The id of the item","type":"integer"},"name":{"title":"The name of the item","type":"string"},"price":{"title":"The price of the item","type":"number"},"image":{"title":"The relative link of the item","type":"string"},"description":{"title":"The description of the item","type":"string"},"quantity":{"title":"The amount of this item that the user has in his cart","type":"integer"}}},"UserDet":{"title":"UserDet","required":["role","email","name"],"type":"object","properties":{"role":{"title":"The role of the user. Either MERCHANT or CUSTOMER","type":"string"},"email":{"title":"The email of the user.","type":"string"},"name":{"title":"The name of the user. [FirstName LastName]","type":"string"}},"example":{"role":"MERCHANT","email":"chipotle@gmail.com","name":"Chipotle Store"}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"type":"string"}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/auth"}}}}}}